fixed pull request
