+++ src/main/java/frc/robot/subsystems/drive/SwerveModule.java
@@ -1,7 +1,5 @@
 package frc.robot.subsystems.drive;
 
-import com.revrobotics.spark.SparkBase.ControlType;
-
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
